requirements.txt
mysql-connector-python==8.0.29
reportlab==3.6.12
Pillow==9.1.0
CustomTkinter==5.1.3
ttkthemes==3.2.2
tkcalendar==1.6.1
